tests/webgl/conformance/uniforms/null-uniform-location.html

<!--

/*
** Copyright (c) 2012 The Khronos Group Inc.
**
** Permission is hereby granted, free of charge, to any person obtaining a
** copy of this software and/or associated documentation files (the
** "Materials"), to deal in the Materials without restriction, including
** without limitation the rights to use, copy, modify, merge, publish,
** distribute, sublicense, and/or sell copies of the Materials, and to
** permit persons to whom the Materials are furnished to do so, subject to
** the following conditions:
**
** The above copyright notice and this permission notice shall be included
** in all copies or substantial portions of the Materials.
**
** THE MATERIALS ARE PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
** MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
** CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
** MATERIALS OR THE USE OR OTHER DEALINGS IN THE MATERIALS.
*/

-->

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../js/js-test-pre.js"></script>
<script src="../../js/webgl-test-utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<script>
"use strict";
description("Tests calling the various uniform[Matrix]* APIs with a null uniform location");

var wtu = WebGLTestUtils;
var gl = wtu.create3DContext();
var program = wtu.loadStandardProgram(gl);

wtu.glErrorShouldBe(gl, gl.NO_ERROR);
shouldBeUndefined("gl.useProgram(program)");
var floatArray = new Float32Array([1, 2, 3, 4]);
var intArray = new Int32Array([1, 2, 3, 4]);

function callUniformFunction(name) {
  var isArrayVariant = (name.charAt(name.length - 1) == 'v');
  var isMatrix = (name.indexOf("Matrix") != -1);
  var isFloat =
      (name.charAt(name.length - 1) == 'f' ||
       name.charAt(name.length - 2) == 'f');
  var sizeIndex = (isArrayVariant ? name.length - 3 : name.length - 2);
  var size = parseInt(name.substring(sizeIndex, sizeIndex + 1));
  // Initialize argument list with null uniform location
  var args = [ null ];
  if (isArrayVariant) {
    // Call variant which takes values as array
    if (isMatrix) {
      size = size * size;
      args.push(false);
    }
    var array = (isFloat ? new Float32Array(size) : new Int32Array(size));
    for (var i = 0; i < size; i++) {
      array[i] = i;
    }
    args.push(array);
  } else {
    // Call variant which takes values as parameters
    for (var i = 0; i < size; i++) {
      args.push(i);
    }
  }
  var func = gl[name];
  return func.apply(gl, args);
}

var funcs = [ "uniform1f", "uniform1fv", "uniform1i", "uniform1iv",
              "uniform2f", "uniform2fv", "uniform2i", "uniform2iv",
              "uniform3f", "uniform3fv", "uniform3i", "uniform3iv",
              "uniform4f", "uniform4fv", "uniform4i", "uniform4iv",
              "uniformMatrix2fv", "uniformMatrix3fv", "uniformMatrix4fv" ];
var callString;

for (var i = 0; i < funcs.length; i++) {
  callString = "callUniformFunction('" + funcs[i] + "')";
  shouldBeUndefined(callString);
  wtu.glErrorShouldBe(gl, gl.NO_ERROR);
}

var successfullyParsed = true;
</script>
<script src="../../js/js-test-post.js"></script>

</body>
</html>
